+++ sfm/__init__.py
@@ -1,0 +1,5 @@
+import sys, os
+from .SFM import SFM
+from .reconstruction import Reconstruction
+
+sys.path.insert(0, os.path.dirname(os.path.realpath(__file__)))
