+++ sfm/reconstruction.py
